Rename dashboard factory namespace
