+++ gms/factories/src/main/java/com/linkedin/gms/factory/dashboard/ChartSearchDaoFactory.java
@@ -1,9 +1,8 @@
-package com.linkedin.dashboard.factory;
+package com.linkedin.gms.factory.dashboard;
 
-
-import com.linkedin.metadata.configs.DashboardSearchConfig;
+import com.linkedin.metadata.configs.ChartSearchConfig;
 import com.linkedin.metadata.dao.search.ESSearchDAO;
-import com.linkedin.metadata.search.DashboardDocument;
+import com.linkedin.metadata.search.ChartDocument;
 import javax.annotation.Nonnull;
 import org.elasticsearch.client.RestHighLevelClient;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,15 +13,15 @@ import org.springframework.context.annotation.DependsOn;
 
 
 @Configuration
-public class DashboardSearchDaoFactory {
+public class ChartSearchDaoFactory {
   @Autowired
   ApplicationContext applicationContext;
 
-  @Bean(name = "dashboardSearchDAO")
+  @Bean(name = "chartSearchDAO")
   @DependsOn({"elasticSearchRestHighLevelClient"})
   @Nonnull
   protected ESSearchDAO createInstance() {
-    return new ESSearchDAO(applicationContext.getBean(RestHighLevelClient.class), DashboardDocument.class,
-        new DashboardSearchConfig());
+    return new ESSearchDAO(applicationContext.getBean(RestHighLevelClient.class), ChartDocument.class,
+        new ChartSearchConfig());
   }
 }
