+++ gms/factories/src/main/java/com/linkedin/gms/factory/dashboard/DashboardSearchDaoFactory.java
@@ -1,8 +1,9 @@
-package com.linkedin.dashboard.factory;
+package com.linkedin.gms.factory.dashboard;
 
-import com.linkedin.metadata.configs.ChartSearchConfig;
+
+import com.linkedin.metadata.configs.DashboardSearchConfig;
 import com.linkedin.metadata.dao.search.ESSearchDAO;
-import com.linkedin.metadata.search.ChartDocument;
+import com.linkedin.metadata.search.DashboardDocument;
 import javax.annotation.Nonnull;
 import org.elasticsearch.client.RestHighLevelClient;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,15 +14,15 @@ import org.springframework.context.annotation.DependsOn;
 
 
 @Configuration
-public class ChartSearchDaoFactory {
+public class DashboardSearchDaoFactory {
   @Autowired
   ApplicationContext applicationContext;
 
-  @Bean(name = "chartSearchDAO")
+  @Bean(name = "dashboardSearchDAO")
   @DependsOn({"elasticSearchRestHighLevelClient"})
   @Nonnull
   protected ESSearchDAO createInstance() {
-    return new ESSearchDAO(applicationContext.getBean(RestHighLevelClient.class), ChartDocument.class,
-        new ChartSearchConfig());
+    return new ESSearchDAO(applicationContext.getBean(RestHighLevelClient.class), DashboardDocument.class,
+        new DashboardSearchConfig());
   }
 }
